SARIF compatibility
* Refactoring
