+++ save-plugins/warn-plugin/src/commonMain/kotlin/org/cqfn/save/plugin/warn/WarnPlugin.kt
@@ -3,7 +3,6 @@ package org.cqfn.save.plugin.warn
 import org.cqfn.save.core.config.ExpectedWarningsFormat
 import org.cqfn.save.core.config.TestConfig
 import org.cqfn.save.core.files.createFile
-import org.cqfn.save.core.files.readFile
 import org.cqfn.save.core.files.readLines
 import org.cqfn.save.core.logging.describe
 import org.cqfn.save.core.logging.logDebug
@@ -12,7 +11,6 @@ import org.cqfn.save.core.plugin.ExtraFlags
 import org.cqfn.save.core.plugin.ExtraFlagsExtractor
 import org.cqfn.save.core.plugin.GeneralConfig
 import org.cqfn.save.core.plugin.Plugin
-import org.cqfn.save.core.plugin.PluginException
 import org.cqfn.save.core.plugin.resolvePlaceholdersFrom
 import org.cqfn.save.core.result.DebugInfo
 import org.cqfn.save.core.result.Fail
@@ -20,26 +18,20 @@ import org.cqfn.save.core.result.TestResult
 import org.cqfn.save.core.utils.ExecutionResult
 import org.cqfn.save.core.utils.ProcessExecutionException
 import org.cqfn.save.core.utils.ProcessTimeoutException
-import org.cqfn.save.plugin.warn.sarif.adjustToCommonRoot
-import org.cqfn.save.plugin.warn.sarif.findAncestorDirContainingFile
-import org.cqfn.save.plugin.warn.sarif.toWarnings
-import org.cqfn.save.plugin.warn.sarif.topmostTestDirectory
 import org.cqfn.save.plugin.warn.utils.ResultsChecker
 import org.cqfn.save.plugin.warn.utils.Warning
+import org.cqfn.save.plugin.warn.utils.collectWarningsFromSarif
 import org.cqfn.save.plugin.warn.utils.collectionMultilineWarnings
 import org.cqfn.save.plugin.warn.utils.collectionSingleWarnings
 import org.cqfn.save.plugin.warn.utils.extractWarning
 import org.cqfn.save.plugin.warn.utils.getLineNumber
 
-import io.github.detekt.sarif4k.SarifSchema210
 import okio.FileNotFoundException
 import okio.FileSystem
 import okio.Path
 import okio.Path.Companion.toPath
 
 import kotlin.random.Random
-import kotlinx.serialization.decodeFromString
-import kotlinx.serialization.json.Json
 
 private typealias WarningMap = Map<String, List<Warning>>
 
@@ -267,27 +259,21 @@ class WarnPlugin(
         generalConfig: GeneralConfig,
         originalPaths: List<Path>,
         originalPath: Path,
-    ): List<Warning> = if (warnPluginConfig.expectedWarningsFormat == ExpectedWarningsFormat.SARIF) {
-        val sarifFileName = warnPluginConfig.expectedWarningsFileName!!
-        val sarif = fs.findAncestorDirContainingFile(originalPath, sarifFileName)?.let { it / sarifFileName }
-            ?: throw PluginException(
-                "Could not find SARIF file with expected warnings for file $this. " +
-                        "Please check if correct `expectedWarningsFormat` is set and if the file is present and called `$sarifFileName`."
-            )
-        val topmostTestDirectory = fs.topmostTestDirectory(originalPath)
-        Json.decodeFromString<SarifSchema210>(
-            fs.readFile(sarif)
+    ): List<Warning> = when {
+        warnPluginConfig.expectedWarningsFormat == ExpectedWarningsFormat.SARIF -> collectWarningsFromSarif(
+            warnPluginConfig,
+            originalPath,
+            originalPaths,
+            fs,
+            this
         )
-            .toWarnings(topmostTestDirectory, originalPaths.adjustToCommonRoot(topmostTestDirectory))
-    } else if (generalConfig.expectedWarningsEndPattern != null) {
-        collectionMultilineWarnings(
+        generalConfig.expectedWarningsEndPattern != null -> collectionMultilineWarnings(
             warnPluginConfig,
             generalConfig,
             fs.readLines(this),
             this,
         )
-    } else {
-        collectionSingleWarnings(
+        else -> collectionSingleWarnings(
             warnPluginConfig,
             generalConfig,
             fs.readLines(this),
