+++ save-plugins/warn-plugin/src/commonMain/kotlin/org/cqfn/save/plugin/warn/utils/WarningsExtraction.kt
@@ -5,10 +5,21 @@
 
 package org.cqfn.save.plugin.warn.utils
 
+import org.cqfn.save.core.files.readFile
 import org.cqfn.save.core.plugin.GeneralConfig
+import org.cqfn.save.core.plugin.PluginException
 import org.cqfn.save.plugin.warn.WarnPluginConfig
+import org.cqfn.save.plugin.warn.sarif.adjustToCommonRoot
+import org.cqfn.save.plugin.warn.sarif.findAncestorDirContainingFile
+import org.cqfn.save.plugin.warn.sarif.toWarnings
+import org.cqfn.save.plugin.warn.sarif.topmostTestDirectory
 
+import io.github.detekt.sarif4k.SarifSchema210
+import okio.FileSystem
 import okio.Path
+
+import kotlinx.serialization.decodeFromString
+import kotlinx.serialization.json.Json
 
 /**
  * @param warnPluginConfig
@@ -84,3 +95,32 @@ internal fun collectionSingleWarnings(
 }
     .filterNotNull()
     .sortedBy { warn -> warn.message }
+
+/**
+ * @param warnPluginConfig
+ * @param originalPath
+ * @param originalPaths
+ * @param fs
+ * @param file
+ * @return a list of warnings extracted from SARIF file for test [file]
+ * @throws PluginException
+ */
+internal fun collectWarningsFromSarif(
+    warnPluginConfig: WarnPluginConfig,
+    originalPath: Path,
+    originalPaths: List<Path>,
+    fs: FileSystem,
+    file: Path,
+): List<Warning> {
+    val sarifFileName = warnPluginConfig.expectedWarningsFileName!!
+    val sarif = fs.findAncestorDirContainingFile(originalPath, sarifFileName)?.let { it / sarifFileName }
+        ?: throw PluginException(
+            "Could not find SARIF file with expected warnings for file $file. " +
+                    "Please check if correct `expectedWarningsFormat` is set and if the file is present and called `$sarifFileName`."
+        )
+    val topmostTestDirectory = fs.topmostTestDirectory(originalPath)
+    return Json.decodeFromString<SarifSchema210>(
+        fs.readFile(sarif)
+    )
+        .toWarnings(topmostTestDirectory, originalPaths.adjustToCommonRoot(topmostTestDirectory))
+}
